split charts by types assigned in array
